Extract and reuse EventItem
