+++ packages/extension/src/views/Events/EventItem/EventItem.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import styled from 'styled-components'
-import { Colors } from '../../design'
-import type { Event } from '../../providers/events/State'
+import { Colors } from '../../../design'
+import type { Event } from '../../../providers/events/State'
 import { Badge } from './Badge'
 import { getEventLabel } from './getEventLabel'
 import { getNetworkColor } from './getNetworkColor'
@@ -9,38 +9,24 @@ import { Latency } from './Latency'
 
 interface Props {
   event: Event
-  selected: boolean
-  onSelect: (event: Event) => void
 }
 
-export const EventListItem = React.memo(function ({ event, selected, onSelect }: Props) {
+export function EventItem({ event }: Props) {
   const networkColor = getNetworkColor(event)
   return (
-    <ListItem className={selected ? 'selected' : ''} onClick={() => onSelect(event)}>
+    <Wrapper>
       <Time>{event.time}</Time>
       <NetworkIndicator style={{ backgroundColor: networkColor }} />
       <Badge event={event} />
       <Label>{getEventLabel(event)}</Label>
       <Latency event={event} />
-    </ListItem>
+    </Wrapper>
   )
-})
+}
 
-const ListItem = styled.li`
+const Wrapper = styled.div`
   display: flex;
   align-items: center;
-  border: 1px solid ${Colors.Background};
-  border-bottom-color: ${Colors.Border};
-  cursor: pointer;
-
-  &:hover {
-    background-color: ${Colors.Hover};
-  }
-
-  &.selected {
-    background-color: ${Colors.Selected};
-    border-color: ${Colors.SelectedBorder};
-  }
 `
 
 const Time = styled.div`
