+++ packages/extension/src/views/Events/EventItem/getNetworkColor.tsx
@@ -1,5 +1,5 @@
-import { Colors } from '../../design'
-import type { Event } from '../../providers/events/State'
+import { Colors } from '../../../design'
+import type { Event } from '../../../providers/events/State'
 
 export function getNetworkColor(event: Event) {
   if (event.type === 'INIT' || event.type === 'NETWORK_DISCONNECTED' || !event.network) {
